Use `false` to signal invalid token, for uniformity.
Previously the `validateClient` and `grantToken` options would signal failure by calling back with `false`, and only calling back with an error in exceptional situations. However, `authenticateToken` relied on calling back with a `restify.UnauthorizedError`. This has been changed so that calling back with `false` in `authenticateToken` acts just as it does in the other two, giving the default not-authenticated behavior.