If a bearer token is supplied for non-authenticated resources, process anyway.
Previously, if `authenticatedRequestPredicate` returned false, but a bearer token was provided, nothing would be done. This was problematic for resources that could have different states depending on whether or not a token was present, since they were either "authenticated" and thus *required* a token, or "public," and thus nothing happened when a token was provided.

Now, `authenticatedRequestPredicate` has been renamed to `requiresTokenPredicate`. And if a token is supplied to a resource that does not require one, Restify-OAuth2 still does its processing, validating the token and assigning the request's `username` property. This allows resources to switch on the presence of the `username` property to tailor their output, and also gives clients useful error messages if they supply invalid tokens to such resources.