Call `next()` in all cases.
Fixes #16. Not only were we failing to call next on certain successes, as in that issue, but we were also doing `res.send(error)` sometimes instead of `next(error)`. Even worse, the tests were not correctly differentiating between the different possible `next` calls (token endpoint vs. plugin handler), so writing a correct test for this was hard.